make roman a library
