+++ roman_test.go
@@ -1,8 +1,7 @@
-package main
+package roman
 
 import (
 	"fmt"
-	"strings"
 	"testing"
 )
 
@@ -68,79 +67,4 @@ func TestRomanNumerals(t *testing.T) {
 			}
 		})
 	}
-}
-
-type RomanNumeral struct {
-	Value  int
-	Symbol string
-}
-
-type RomanNumerals []RomanNumeral
-
-func (r RomanNumerals) ValueOf(symbols ...byte) int {
-	symbol := string(symbols)
-	for _, s := range r {
-		if s.Symbol == symbol {
-			return s.Value
-		}
-	}
-	return 0
-}
-
-var allRomanNumerals = RomanNumerals{
-	{1000, "M"},
-	{900, "CM"},
-	{500, "D"},
-	{400, "CD"},
-	{100, "C"},
-	{90, "XC"},
-	{50, "L"},
-	{40, "XL"},
-	{10, "X"},
-	{9, "IX"},
-	{5, "V"},
-	{4, "IV"},
-	{1, "I"},
-}
-
-func ConvertToRoman(arabic int) string {
-
-	var result strings.Builder
-
-	for _, numeral := range allRomanNumerals {
-		for arabic >= numeral.Value {
-			result.WriteString(numeral.Symbol)
-			arabic -= numeral.Value
-		}
-	}
-
-	return result.String()
-}
-
-func ConvertToArabic(roman string) int {
-
-	total := 0
-
-	for i := 0; i < len(roman); i++ {
-
-		symbol := roman[i]
-
-		if couldBeSubtractive(i, symbol, roman) {
-			if value := allRomanNumerals.ValueOf(symbol, roman[i+1]); value != 0 {
-				total += value
-				i++
-			} else {
-				total += allRomanNumerals.ValueOf(symbol)
-			}
-		} else {
-			total += allRomanNumerals.ValueOf(symbol)
-		}
-	}
-
-	return total
-}
-
-func couldBeSubtractive(index int, currentSymbol uint8, roman string) bool {
-	isSubtractive := currentSymbol == 'I' || currentSymbol == 'X' || currentSymbol == 'C'
-	return index+1 < len(roman) && isSubtractive
 }
